Refactor: PD-90 Account 섹션 및 스타일 분리
