+++ src/pages/sign-up/style/variants.ts
@@ -1,0 +1,3 @@
+import { cva } from 'class-variance-authority'
+
+export const field = cva('flex flex-col gap-[6px] not-last:mb-6')
