+++ src/pages/sign-up/ui/account/variants.ts
@@ -1,0 +1,5 @@
+import { cva } from 'class-variance-authority'
+
+export const textFieldWrapper = cva('flex gap-2')
+
+export const passwordFieldWrapper = cva('flex flex-col gap-2')
