+++ src/pages/sign-up/ui/sign-up-page/variants.ts
@@ -12,12 +12,6 @@ export const goToSignIn = cva(
   'body-large flex items-center gap-1 text-gray-700',
 )
 
-export const field = cva('flex flex-col gap-[6px] not-last:mb-6')
-
-export const textFieldWrapper = cva('flex gap-2')
-
-export const passwordFieldWrapper = cva('flex flex-col gap-2')
-
 export const radioFieldWrapper = cva('flex gap-[30px]')
 
 export const checkbox = cva('mb-4 pl-[15px]')
